voxel_proxy/requirements.txt
Autodesk_Maya
Bifrost 2.0.5 +
Python 3
MeshLab 2020.03 (with meshlabserver.exe)